+++ benchmarks/src/main/scala/cats/effect/benchmarks/RedeemBenchmark.scala
@@ -17,7 +17,7 @@
 package cats.effect.benchmarks
 
 import cats.effect.IO
-import cats.effect.unsafe.implicits.global
+import cats.effect.unsafe.IORuntime.Implicits.global
 
 import org.openjdk.jmh.annotations._
