+++ docs/datatypes/dequeue.md
@@ -53,7 +53,7 @@ import cats.{Contravariant, Functor}
 import cats.implicits._
 import cats.effect._
 import cats.effect.std.{Dequeue, DequeueSource, DequeueSink}
-import cats.effect.unsafe.implicits.global
+import cats.effect.unsafe.IORuntime.Implicits.global
 
 def covariant(list: List[Int]): IO[List[Long]] = (
   for {
